requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
flask-cors==3.0.10
opencv-python-headless==4.5.5.62
numpy==1.21.5
Pillow==8.4.0
networkx
scipy
tensorboardX
matplotlib